requirements.txt
ccxt
aiosqlite
pandas
pandas_ta
pyyaml
matplotlib
streamlit
pydantic
aiohttp
